% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_tabu_bin_bisect.R
\name{z_local_search}
\alias{z_local_search}
\title{Finding the Value of z Statistics for Dynamic Significance Level}
\usage{
z_local_search(level_of_significance, tol, max_iteration,
li = -7, ui = 7, is_echo = FALSE)
}
\arguments{
\item{level_of_significance}{a value between 0 and 1}

\item{tol}{the tolerance quantity, usually 0.001}

\item{max_iteration}{the maximum iteration as termination criteria of searching}

\item{li}{the lower bound of the searching}

\item{ui}{the upper bound of the search space}

\item{is_echo}{a boolean value which is by default, it's FALSE.}
}
\value{
a number (exclusive) or data frame.
}
\description{
\code{z_local_search} is used to find the critical value of z
statistics at level of significance between 0 and 1
}
\details{
Function z_local_search is used to find the critical
value of z statistics at level of significance between 0 and 1.
The critical value is used in the construction of interval to
estimate a certain parameter, and it's used in hypothesis testing
as well.

Level of significance is correspond to the confidence limits, if
confidence limits is known, then the level of significance can be
obtained by subtracting the value of confidence limits from 1.
Tolerance quantity is used to find this critical point statistics
correct up to i-th decimal places, usually it's set to be 0.001.
The tolerance quantity is used as terminaton criteria too.
Maximum iteration max_itear is used to the termination criteria
too. The li and ui are, respectively, the lower bound and upper
bound of the searching. For z-statistics, by default, the value
of li and ui are, respectively, -7 and 7. The value of is_echo
is set ot be FALSE by default, since most of people only want to
obtained this critical value. If is_echo is TRUE, then the
function will return an iteration table as data frame.
}
\note{
In order to check wether the return value is indeed, the critical value,
use pnorm function
}
\examples{
# Given the level of significance is 0.025,
# then the critical value for z at this level of
# significance is computed as follows.
z_local_search(level_of_significance = 0.025, tol = 0.001,
max_iteration = 50
) #[1] 1.960062
pnorm(1.960062) #[1] 0.9750057
1-pnorm(1.960062) #[1] 0.02599427 ~= 0.025

z_local_search(level_of_significance = 0.025, tol = 0.001,
max_iteration = 50, is_echo = TRUE) # return a data frame
}
\references{
Howard, J.P. (2017). Computational Methods for Numerical Analysis with R (1st ed.).
Chapman and Hall/CRC. https://doi.org/10.1201/9781315120195

Laguna M. (2018) Tabu Search. In: Martí R., Pardalos P., Resende M. (eds) Handbook
of Heuristics. Springer, Cham. https://doi.org/10.1007/978-3-319-07124-4_24

Ronald E. Walpole, Raymond H. Myers, Sharon L. Myers, and Keying Ye. Probability
& Statistics for Engineers & Scientists, Nineth Global edition (9th glob. ed.).
Pearson.

Thomas H. Cormen, Charles E. Leiserson, Ronald L. Rivest, and Clifford Stein. 2009.
Introduction to Algorithms, Third Edition (3rd. ed.). The MIT Press.
}
\seealso{
If one needs critical value of t, then use t_local_search
function. Similarly for chisq_local_search, and f_local_search
}
